renames 'user-profile' folder to 'profile'
