+++ src/components/pages/profile/Favorites.js
