+++ src/components/pages/profile/profile.scss
